% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{clockplot}
\alias{clockplot}
\title{Draw 2 clock plots to show the true response values and predicted response
values}
\usage{
clockplot(comptab, title, jitterval = TRUE, colorful = FALSE)
}
\arguments{
\item{comptab}{A matrix or data.frame with the predicted and true response
values. Each row is a sample, and a column named "Prediction" should be
used to record the model predicted values, while a column named "True"
should be used to record the corresponding true values. They should be
numeric values. In the results generated by the function
\code{singleselection}, \code{singlebalance} or \code{crosstraining}, the
slots named "traincomp" and "testcomp" can be used directly as the input
of this parameter.}

\item{title}{The title of the plot.}

\item{colorful}{Whether the pointers in the clock plots should be colored
using gradually changed colors to reflect the density of the pointers in
the plots. If it is set as TRUE and the total pointer number is greater
than 50, the density color will be used. Default is FALSE.}

\item{jitter}{In some cases, the true values of the samples have a low
precision, which makes many samples have the same true value and cannot be
distinguished in the plot. This parameter can be used to introduce some
random jitters into the true values to distinguish them in the plot.}
}
\value{
Two clock plots with one showing the true sample response values and
the other showing the predicted response values. In each clock, a pointer
represents a sample and the scale around the dial is the scale of the true
or predicted response.
}
\description{
Draw 2 clock plots to show the true response values and predicted response
values.
}
