% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{dividedat}
\alias{dividedat}
\title{Divide a dataset into training dataset and testing dataset}
\usage{
dividedat(i = 1, fold = 0.9, totalpd, totalvar)
}
\arguments{
\item{i}{Random seed need to be set for dividing the dataset randomly.
Default is 1.}

\item{fold}{A float number between 0 and 1, indicating the proportion of
total samples that will be included in the training dataset, so that the
proportion of total samples that will be included in the testing dataset
is \code{1 - fold}. Default is 0.9.}

\item{totalpd}{A data.frame indicating the meta information of all samples.
Each row represents one sample and each column represents one feature. Two
columns must be included in this data.frame, one is a column named
"sampleid", which records the unique IDs of samples, and the other is a
column named "Response", which records the values of the response variable
that will be used to construct the downstream machine learning model, and
this column can be a factor, or record numeric values. If it is a factor,
it will be converted to numeric values.}

\item{totalvar}{A matrix recording the values of candidate features for all
samples. Each row represents one sample and each column represents one
feature. The column names are the feature names while the row names should
be sample IDs. Each column records a series of numeric values.}
}
\value{
A list with tow sub-lists. One is a sub-list named "trainset"
recording the information of samples that have been attributed to the
training dataset. It contains 3 slots. The one named "innertrainvar" has
the same structure as the data transferred to the parameter
\code{totalvar}, another one named "innertrainpd" including the meta
information for the training samples similar to the data transferred to
\code{totalpd}. The third one named "innertrainresponse" is a one-column
frame with the value of the response variable as its only column and the
row names are training sample IDs. The second sub-list is named "testset",
and it contains 2 slots named "innertestvar" and "innertestpd" that record
the candidate feature values and meta information for samples in the
testing dataset.
}
\description{
Divide a dataset into training dataset and testing dataset randomly.
}
