% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{ensemblepredict}
\alias{ensemblepredict}
\title{Calculate model predicted response values for the input data}
\usage{
ensemblepredict(balanceres, vardat)
}
\arguments{
\item{balanceres}{The model trained. The result generated by the function
\code{singlenet} or \code{singlebalance} can be directly used for this
parameter, while for the function \code{crosstraining}, the element in its
sub-list named "modellist" could also be used for this parameter.}

\item{vardat}{The input data whose response values need to be predicted by
the model. Each row is a sample and each column is a feature (e.g. a
methylation probe). The values in \code{vardat} should be numeric values.}
}
\value{
A one-column matrix containing the model predicted response values.
Column name of this matrix is "Prediction" and each row of it represents a
sample.
}
\description{
Predict response values for the input data using a trained elastic net model
or ensemble model.
}
