% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{extractprobes}
\alias{extractprobes}
\title{Extract features selected}
\usage{
extractprobes(res)
}
\arguments{
\item{res}{Results generated by the function \code{singleselection},
\code{singlebalance}, or \code{crosstraining}.}
}
\value{
A data.frame with two columns. One is named "Probe" containing the
features extracted from the single model trained by the function
\code{singleselection} or \code{singlebalance}, or from all the models
trained by the function \code{crosstraining}. The other is a column named
"Coeffsum", and for each feature, its value is the sum of its regression
coefficients in all the models contained in \code{res}.
}
\description{
Extract features selected from the models.
}
