% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{heatmapplot}
\alias{heatmapplot}
\title{Draw a heatmap to show the features and the response variable}
\usage{
heatmapplot(orivar, comptab, featurenames = NULL, title)
}
\arguments{
\item{orivar}{A matrix with the feature values need to be displayed in the
heatmap. Each row represents one sample and each column represents one
feature. The column names are the feature names while the row names should
be sample IDs. Each column records a series of numeric values.}

\item{comptab}{A matrix or data.frame with the predicted and true response
values. Each row is a sample, and a column named "Prediction" should be
used to record the model predicted values, while a column named "True"
should be used to record the corresponding true values. They should be
numeric values. The row names should be sample IDs same as that in the
\code{orivar} matrix. In the results generated by the function
\code{singleselection}, \code{singlebalance} or \code{crosstraining}, the
slots named "traincomp" and "testcomp" can be used directly as the input
of this parameter.}

\item{featurenames}{Name of features used by the model to predict response
that will be shown by the heatmap. Default is NULL, which means all the
features included in the \code{orivar} matrix will be shown. If want to
show the features (the rows in the heatmap) with a specific order, this
parameter need to be given as a vector with the ordered feature names.}

\item{title}{The title of the plot.}
}
\value{
Three heatmaps will be shown. For all of them, a row is a feature,
and a column is a sample. The first heatmap has the samples ordered by
their true response values. The second one has the samples ordered by
their predicted response values. While the features are ordered according
to the parameter \code{featurenames}. The third heatmap has both the
samples and the features hierarchically clustered using a "complete"
method.
}
\description{
Draw a heatmap to show the features and the predicted and true response
values.
}
