% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{probeannotation}
\alias{probeannotation}
\title{Annotate the Illumina methylation probes}
\usage{
probeannotation(platform = 450, finalprobes)
}
\arguments{
\item{platform}{The platform of the probes need to be annotated. Can be set
as 27 (for 27k platform), 450 (for 450k platform), or 850 (for EPIC
platform). Default is 450.}

\item{finalprobes}{The names of the probes need to be annotated.}
}
\value{
A data.frame with the probe annotation information, including the
genomic loci of the probes and their related CpG islands and genes.
}
\description{
Annotate the Illumina methylation probes
}
