% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{probetodmr}
\alias{probetodmr}
\title{Summarize the beta values of probes to DNA methylation regions (DMRs)}
\usage{
probetodmr(betadat, platform = 450, maxgap = 300, TSSradius = 1500)
}
\arguments{
\item{betadat}{A matrix recording the beta values of methylation probes for
samples. Each column represents one sample and each row represents one
probe. The row names are the probe names while the column names should be
sample IDs.}

\item{platform}{The platform of the probes. Can be set as 27 (for 27k
platform), 450 (for 450k platform), or 850 (for EPIC platform). Default is
450.}

\item{maxgap}{An integer indicating the cutoff of probe-probe distance when
clustering the probes into DNA methylation regions (DMRs). If the distance
between 2 neighbor probes is less than this cutoff, they will be clustered
into the same DMR. Default is 300.}

\item{TSSradius}{An integer defining the TSS region that will be considered
when mapping the genes to DMRs. If an DMR overlaps with a gene region from
\code{TSSradius} bp upstream to \code{TSSradius} bp downstream of the TSS,
this gene will be attributed to this DMR. In case when one gene region
overlaps with more than one DMR, it will be attributed to the DMR with the
closest distance to it, so that the genes covered by the DMRs can be
annotated. Default value is 1500.}
}
\value{
A list with 3 slots. The slot named "betadat" is a matrix recording
the DMR beta values of samples. The slot named "dmrprobemapping" is a
data.frame recording the probes covered by each DMR and the coordinates
and other information of the DMRs and probes. The slot "dmrgenemapping" is
a data.frame recording the genes whose TSS regions are covered by each
DMR.
}
\description{
Cluster the probes into DNA methylation regions (DMRs) and calculate the
beta values of the DMRs via averaging the probe beta values within it.
}
