% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{resamplebin}
\alias{resamplebin}
\title{Generate data with balanced response distribution from an imbalanced one}
\usage{
resamplebin(
  resrange,
  binwith = 1,
  minwith = NULL,
  sampleseed = 1,
  samplevar,
  upsampling = "SMOTE",
  plotting = TRUE,
  balanceidx = FALSE
)
}
\arguments{
\item{resrange}{A data.frame with meta information of the samples. Two
columns must be contained in it. One is the column "sampleid", the other
is the column "Response", which records the response variable values of
the samples. The response values should be numeric values.}

\item{binwith}{The start value of the bin width used to divide the response
values into different bins and calculate their distribution. Its value
will be increased a bit automatically for each recursion loop conducted by
this function. The default value of this parameter is 1.}

\item{minwith}{The increased value on \code{binwith} for each recursion loop
of this function. Its default value is \code{NULL}, meaning it is the same
as the original \code{binwith} value.}

\item{sampleseed}{The random seed used in this function to do sampling. Its
default value is 1.}

\item{samplevar}{A matrix recording the values of features for all samples.
Each row represents one sample and each column represents one feature. The
column names are the feature names while the row names should be sample
IDs same as the ones recorded by the "sampleid" column in \code{resrange}.
Each column records a series of numeric values.}

\item{upsampling}{The method to do over-sampling for the samples with a low
response distribution density. The over-sampling will make the density of
these samples increase to the level of other samples. Its default value is
"SMOTE", which means that the SMOTE (Synthetic Minority Over-sampling
Technique) method will be used. The value of this parameter can also be
"boot", which means bootstrapping will be used, but it is not recommended
because it is more likely to bring over-fitting problems.}

\item{plotting}{Whether or not to plot the sample response distribution
before and after the adjustment by this function. Default value is TRUE.}

\item{balanceidx}{Whether or not to calculate the balance index of the
sample response distribution before the adjustment, which is the variance
of the distribution density values around their mean.}
}
\value{
A list with several elements. One is a data.frame recording the
sample meta data of the new dataset generated by this function. The
distribution of the response variable will be more balanced. The other
element is the feature value matrix for the new samples generated. Its
format is the same as the one for the parameter \code{samplevar}. If the
parameter \code{balanceidx} is set as TRUE, also a list element containing
the balance index before distribution adjustment will be returned.
}
\description{
Use bootstrapping or SMOTE method to sample the original samples and
generate new data with a balanced response distribution.
}
