% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{residualcomp}
\alias{residualcomp}
\title{Plot the sample response-sample residual relationship for multiple models}
\usage{
residualcomp(comptabs, title, methodnames)
}
\arguments{
\item{comptabs}{A list with more than one matrix or data.frame as elements,
Each matrix or data.frame corresponds to one model and should contain the
predicted and true response values of that model. Each row is a sample,
and a column named "Prediction" should be used to record the model
predicted values, while a column named "True" should be used to record the
corresponding true values. They should be numeric values. In the results
generated by the function \code{singleselection}, \code{singlebalance} or
\code{crosstraining}, the slots named "traincomp" and "testcomp" can be
used as these matrices or data.frames.}

\item{title}{The title of the plot.}

\item{methodnames}{A vector containing the names of the models. It
corresponds to the list of \code{comptabs} and its length should be the
same as the list length.}
}
\value{
A plot showing the true sample response-sample residual relationship
for multiple models.
}
\description{
Draw a scatter plot to show the sample response-sample residual relationship
for multiple models.
}
