% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{residualplot}
\alias{residualplot}
\title{Plot the true sample response-sample residual relationship}
\usage{
residualplot(comptab, featurenum = NULL, title, colorful = FALSE)
}
\arguments{
\item{comptab}{A matrix or data.frame with the predicted and true response
values. Each row is a sample, and a column named "Prediction" should be
used to record the model predicted values, while a column named "True"
should be used to record the corresponding true values. They should be
numeric values. In the results generated by the function
\code{singleselection}, \code{singlebalance} or \code{crosstraining}, the
slots named "traincomp" and "testcomp" can be used directly as the input
of this parameter.}

\item{featurenum}{Number of features used by the model to predict response.
Will be shown in the subtitle of the plot. Default is NULL, which means
this information will not be shown in the plot.}

\item{title}{The title of the plot.}

\item{colorful}{Whether the dots in the plot should be colored using
gradually changed colors to reflect the density of the dots in the plot.
If it is set as TRUE and the total dot number is greater than 50, the
density color will be used. Default is FALSE.}
}
\value{
A scatter plot showing the true sample response-sample residual
relationship.
}
\description{
Draw a scatter plot to show the relationship between the true response
values and the sample residuals of the model.
}
