% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{simpleboot}
\alias{simpleboot}
\title{Do bootstrapping on the samples of a dataset}
\usage{
simpleboot(resrange, sampleseed = 1, samplevar, plotting = TRUE)
}
\arguments{
\item{resrange}{A data.frame with meta information of the samples. Two
columns must be contained in it. One is the column "sampleid", the other
is the column "Response", which records the response variable values of
the samples. The response values should be numeric values.}

\item{sampleseed}{The random seed used in this function to do sampling. Its
default value is 1.}

\item{samplevar}{A matrix recording the values of features for all samples.
Each row represents one sample and each column represents one feature. The
column names are the feature names while the row names should be sample
IDs same as the ones recorded by the "sampleid" column in \code{resrange}.
Each column records a series of numeric values.}

\item{plotting}{Whether or not to plot the sample response distribution
before and after the bootstrapping by this function. Default value is
TRUE.}
}
\value{
A list with 2 elements. One is a data.frame recording the sample
meta data of the bootstrapped dataset generated by this function. The
other element is the feature value matrix for the bootstrapped samples
generated. Its format is the same as the one for the parameter
\code{samplevar}.
}
\description{
Use bootstrapping to sample the original samples and generate new
bootstrapped data.
}
