% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{singleselection}
\alias{singleselection}
\title{Train an elastic net model to fit the response and do feature selection}
\usage{
singleselection(
  orivar,
  oripd,
  fold = 0.9,
  seednum = 1,
  cores = 1,
  alphas = c(0.5),
  errortype = "min",
  localrsquare = FALSE,
  plotting = FALSE,
  balanceidx = FALSE,
  binwith = 1,
  prescreen = FALSE
)
}
\arguments{
\item{orivar}{A matrix recording the values of candidate features for all
samples (training samples + testing samples). Each row represents one
sample and each column represents one feature. The column names are the
feature names while the row names should be sample IDs. Each column
records a series of numeric values.}

\item{oripd}{A data.frame indicating the meta information of all samples
(training samples + testing samples). Each row represents one sample and
each column represents one feature. Two columns must be included in this
data.frame, one is a column named "sampleid", which records the unique IDs
of samples, and the other is a column named "Response", which records the
values of the response variable that will be used to construct the
downstream elastic net model, and this column can be a factor, or record
numeric values. If it is a factor, it will be converted to numeric values.}

\item{fold}{A float number between 0 and 1, indicating the proportion of
total samples that will be included in the training dataset, so that the
proportion of total samples that will be included in the testing dataset
is \code{1 - fold}. Default is 0.9.}

\item{seednum}{Random seed need to be set for dividing the dataset into
training and testing randomly. Default if 1.}

\item{cores}{Number of threads need to be used to parallelize the computing.
Default is 1.}

\item{alphas}{A vector with the candidate elastic net mixing parameters,
which should be between 0 and 1. If the mixing parameter (alpha) is 1, the
regularization is actually a LASSO regularization. If alpha is 0, the
regularization is actually a ridge regularization. Default is 0.5.}

\item{errortype}{The method to automatically choose the regularization
constant lambda and the elastic net mixing parameter alpha. If it is
"min", the lambda and alpha combination giving minimum mean 10-fold cross-
validation error on the training set will be used, while if it is "1ses",
the lambda and alpha combination giving the cross-validation error within
one standard error of the minimum will be used. Default is "min".}

\item{localrsquare}{Whether the local R squares of several small sample
groups within specific response value intervals need to be calculated.
Default is FALSE.}

\item{plotting}{Whether several plots need to be generated to show the
performance of the model, including the scatter plots, heatmap plots, and
residual plots on both training and testing datasets. Default is FALSE.}

\item{balanceidx}{Whether or not to calculate the balance index of the
original training data response distribution.}

\item{binwith}{If the parameter code{balanceidx} is set as TRUE, this
parameter will be needed to set the start value of the bin width to divide
the response variable and to measure this distribution. Its default value
is 1.}

\item{prescreen}{If this parameter is set as TRUE, before the elastic net
model construction, a preliminary screen on the features included in
\code{orivar} will be performed first according to the p-value of their
Pearson correlation coefficients to the response variable, and the ones
with such a p-value less than 0.05 will be selected to train the elastic
net model. Default is FALSE.}
}
\value{
A list with several slots. The slot named "elasticmodel" contains
an S3 object for the final model with the optimal lambda and alpha values.
The slot "modelcoefs" records the regression coefficients for all the
candidate features as well as the intercept of the model. The slots
"traincomp" and "testcomp" are data.frames with the predicted and true
response values for training and testing datasets. The column names are
"Prediction" and "True", the row names are the sample IDs. If the
parameter \code{plotting} is set as TRUE, several plots reflecting the
model performance will also be returned.
}
\description{
Divide the original dataset into training and testing for one time and then
train an elastic net model and do feature selection.
}
